feat: unify central and peripheral codebases and cleanup unnecessary image headers
