+++ boards/shields/nice_view_elemental/CMakeLists.txt
@@ -1,45 +1,25 @@
 if(CONFIG_ZMK_DISPLAY AND CONFIG_NICE_VIEW_WIDGET_STATUS)
-  zephyr_library_include_directories(${CMAKE_SOURCE_DIR}/include)
+    zephyr_library_include_directories(${CMAKE_SOURCE_DIR}/include)
 
-  zephyr_library_sources(
-    ./src/main.c
-    ./src/images/bluetooth_logo_outlined.c
-    ./src/images/bluetooth_logo.c
-    ./src/images/bluetooth_searching.c
-    ./src/images/usb_logo.c
-    ./src/utils/draw_battery.c
-    ./src/utils/draw_bluetooth_logo_outlined.c
-    ./src/utils/draw_bluetooth_logo.c
-    ./src/utils/draw_bluetooth_searching.c
-    ./src/utils/draw_usb_logo.c
-    
-    ./src/images/grid.c
-    ./src/images/grid_inverse.c
-    ./src/images/battery.c
-    ./src/images/battery_bolt.c
-    ./src/images/usb.c
-  )
+    zephyr_library_sources(
+        ./src/fonts/press_start_2p_8.c
+        ./src/fonts/press_start_2p_16.c
+        ./src/fonts/press_start_2p_24.c
 
-  if(NOT CONFIG_ZMK_SPLIT OR CONFIG_ZMK_SPLIT_ROLE_CENTRAL)
-    zephyr_library_sources(
-      ./src/central/initialize_listeners.c
-      ./src/central/render.c
-      # These are not included on the peripheral since they are only used for
-      # displaying the layer name.
-      ./src/fonts/press_start_2p_8.c
-      ./src/fonts/press_start_2p_16.c
-      ./src/fonts/press_start_2p_24.c
+        ./src/images/battery_bolt.c
+        ./src/images/battery.c
+        ./src/images/bluetooth_connected.c
+        ./src/images/bluetooth_disconnected.c
+        ./src/images/bluetooth_searching.c
+        ./src/images/grid_inverse.c
+        ./src/images/grid.c
+        ./src/images/usb.c
+        
+        ./src/utils/draw_battery.c
+        ./src/utils/draw_image.c
+        
+        ./src/initialize_listeners.c
+        ./src/main.c
+        ./src/render.c
     )
-  else()
-    zephyr_library_sources(
-      ./src/peripheral/initialize_listeners.c
-      ./src/images/noise_0.c
-      ./src/images/noise_1.c
-      ./src/images/noise_2.c
-      ./src/images/noise_3.c
-      ./src/images/noise_4.c
-      ./src/images/noise_5.c
-      ./src/peripheral/render.c
-    )
-  endif()
 endif()
