--- boards/shields/nice_view_elemental/include/central/initialize_listeners.h
@@ -1,38 +1,0 @@
-#pragma once
-
-#include "../initialize_listeners.h"
-
-#include <stdbool.h>
-#include <zmk/endpoints.h>
-#include <zmk/keymap.h>
-#include "../utils/draw_battery.h"
-
-struct connectivity_state {
-    struct zmk_endpoint_instance selected_endpoint;
-    int active_profile_index;
-    bool active_profile_connected;
-    bool active_profile_bonded;
-};
-
-struct layer_state {
-    zmk_keymap_layer_index_t index;
-    const char* name;
-};
-
-struct modifiers_state {
-    bool is_shift_active;
-    bool is_ctrl_active;
-    bool is_alt_active;
-    bool is_gui_active;
-};
-
-struct states {
-    struct battery_state battery;
-    struct connectivity_state connectivity;
-    struct layer_state layer;
-    struct modifiers_state modifiers;
-};
-
-extern struct states states;
-
-void initialize_listeners();
