--- boards/shields/nice_view_elemental/include/images/bluetooth_searching.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <lvgl.h>
-
-extern const lv_image_dsc_t bluetooth_searching;
