+++ boards/shields/nice_view_elemental/include/main.h
@@ -51,4 +51,3 @@ extern lv_obj_t* modifiers_canvas;
 extern lv_color_t modifiers_canvas_buffer[];
 
 extern lv_obj_t* image_canvas;
-extern lv_color_t image_canvas_buffer[];
