--- boards/shields/nice_view_elemental/include/peripheral/initialize_listeners.h
@@ -1,19 +1,0 @@
-#pragma once
-
-#include "../initialize_listeners.h"
-
-#include <stdbool.h>
-#include "../utils/draw_battery.h"
-
-struct connectivity_state {
-    bool connected;
-};
-
-struct states {
-    struct battery_state battery;
-    struct connectivity_state connectivity;
-};
-
-extern struct states states;
-
-void initialize_listeners();
