--- boards/shields/nice_view_elemental/include/peripheral/render.h
@@ -1,7 +1,0 @@
-#pragma once
-
-void render_battery();
-void render_connectivity();
-void initialize_animation();
-void start_animation();
-void stop_animation();
