+++ boards/shields/nice_view_elemental/include/render.h
@@ -4,3 +4,6 @@ void render_battery();
 void render_connectivity();
 void render_main();
 void render_modifiers();
+void initialize_animation();
+void start_animation();
+void stop_animation();
