+++ boards/shields/nice_view_elemental/include/utils/draw_image.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <lvgl.h>
+
+void draw_image(const lv_image_dsc_t* image, lv_obj_t* canvas, lv_coord_t x, lv_coord_t y);
