--- boards/shields/nice_view_elemental/include/utils/draw_usb_logo.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <lvgl.h>
-
-void draw_usb_logo(lv_obj_t* canvas, lv_coord_t x, lv_coord_t y);
