+++ boards/shields/nice_view_elemental/src/images/bluetooth_disconnected.c
@@ -1,0 +1,51 @@
+#include <lvgl.h>
+
+#ifndef LV_ATTRIBUTE_MEM_ALIGN
+#define LV_ATTRIBUTE_MEM_ALIGN
+#endif
+
+#ifndef LV_ATTRIBUTE_BLUETOOTH_DISCONNECTED
+#define LV_ATTRIBUTE_BLUETOOTH_DISCONNECTED
+#endif
+
+static const
+LV_ATTRIBUTE_MEM_ALIGN LV_ATTRIBUTE_LARGE_CONST LV_ATTRIBUTE_BLUETOOTH_DISCONNECTED
+uint8_t bluetooth_disconnected_map[] = {
+
+    0xfe,0x00,0xfe,0x00,0xfe,0x00,0xfe,0x00,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0x00,0xfe,0x00,0xfe,0x00,0xfe,0x00,
+    0xfe,0x00,0xfe,0x00,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0x00,0xfe,0x00,
+    0xfe,0x00,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0x00,
+    0xfe,0x00,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0x00,
+    0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,
+    0xfe,0x00,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0x00,
+    0xfe,0x00,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0x00,
+    0xfe,0x00,0xfe,0x00,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0x00,0xfe,0x00,
+    0xfe,0x00,0xfe,0x00,0xfe,0x00,0xfe,0x00,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0x00,0xfe,0x00,0xfe,0x00,0xfe,0x00,
+
+};
+
+const lv_image_dsc_t bluetooth_disconnected = {
+  .header = {
+    .magic = LV_IMAGE_HEADER_MAGIC,
+    .cf = LV_COLOR_FORMAT_AL88,
+    .flags = 0,
+    .w = 14,
+    .h = 19,
+    .stride = 28,
+    .reserved_2 = 0,
+  },
+  .data_size = sizeof(bluetooth_disconnected_map),
+  .data = bluetooth_disconnected_map,
+  .reserved = NULL,
+};
+
