--- boards/shields/nice_view_elemental/src/images/bluetooth_logo_outlined.c
@@ -1,43 +1,0 @@
-#include "../../include/images/bluetooth_logo_outlined.h"
-
-#include <lvgl.h>
-
-#ifndef LV_ATTRIBUTE_MEM_ALIGN
-#define LV_ATTRIBUTE_MEM_ALIGN
-#endif
-
-#ifndef LV_ATTRIBUTE_IMG_BLUETOOTH_LOGO_OUTLINED
-#define LV_ATTRIBUTE_IMG_BLUETOOTH_LOGO_OUTLINED
-#endif
-
-const LV_ATTRIBUTE_MEM_ALIGN LV_ATTRIBUTE_LARGE_CONST LV_ATTRIBUTE_IMG_BLUETOOTH_LOGO_OUTLINED uint8_t bluetooth_logo_outlined_map[] = {
-  0xfe, 0xfe, 0xfe, 0xff, 	/*Color of index 0*/
-  0x00, 0x00, 0x00, 0xff, 	/*Color of index 1*/
-
-  0x0f, 0x00,
-  0x30, 0xc0,
-  0x40, 0x20,
-  0x44, 0x20,
-  0x86, 0x10,
-  0x85, 0x10,
-  0x94, 0x90,
-  0x8d, 0x10,
-  0x86, 0x10,
-  0x8d, 0x10,
-  0x94, 0x90,
-  0x85, 0x10,
-  0x86, 0x10,
-  0x44, 0x20,
-  0x40, 0x20,
-  0x30, 0xc0,
-  0x0f, 0x00,
-};
-
-const lv_image_dsc_t bluetooth_logo_outlined = {
-  .header.cf = LV_COLOR_FORMAT_I1,
-  .header.magic = LV_IMAGE_HEADER_MAGIC,
-  .header.w = 12,
-  .header.h = 17,
-  .data_size = 42,
-  .data = bluetooth_logo_outlined_map,
-};
