+++ boards/shields/nice_view_elemental/src/images/bluetooth_searching.c
@@ -1,43 +1,51 @@
-#include "../../include/images/bluetooth_searching.h"
-
 #include <lvgl.h>
 
 #ifndef LV_ATTRIBUTE_MEM_ALIGN
 #define LV_ATTRIBUTE_MEM_ALIGN
 #endif
 
-#ifndef LV_ATTRIBUTE_IMG_BLUETOOTH_SEARCHING
-#define LV_ATTRIBUTE_IMG_BLUETOOTH_SEARCHING
+#ifndef LV_ATTRIBUTE_BLUETOOTH_SEARCHING
+#define LV_ATTRIBUTE_BLUETOOTH_SEARCHING
 #endif
 
-const LV_ATTRIBUTE_MEM_ALIGN LV_ATTRIBUTE_LARGE_CONST LV_ATTRIBUTE_IMG_BLUETOOTH_SEARCHING uint8_t bluetooth_searching_map[] = {
-  0xfe, 0xfe, 0xfe, 0xff, 	/*Color of index 0*/
-  0x00, 0x00, 0x00, 0xff, 	/*Color of index 1*/
+static const
+LV_ATTRIBUTE_MEM_ALIGN LV_ATTRIBUTE_LARGE_CONST LV_ATTRIBUTE_BLUETOOTH_SEARCHING
+uint8_t bluetooth_searching_map[] = {
 
-  0x0f, 0x00, 
-  0x30, 0xc0, 
-  0x40, 0x20, 
-  0x40, 0x20, 
-  0x82, 0x10, 
-  0x81, 0x10, 
-  0x89, 0x10, 
-  0x84, 0x90, 
-  0x94, 0x90, 
-  0x84, 0x90, 
-  0x89, 0x10, 
-  0x81, 0x10, 
-  0x82, 0x10, 
-  0x40, 0x20, 
-  0x40, 0x20, 
-  0x30, 0xc0, 
-  0x0f, 0x00, 
+    0xfe,0x00,0xfe,0x00,0xfe,0x00,0xfe,0x00,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0x00,0xfe,0x00,0xfe,0x00,0xfe,0x00,
+    0xfe,0x00,0xfe,0x00,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0x00,0xfe,0x00,
+    0xfe,0x00,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0x00,
+    0xfe,0x00,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0x00,
+    0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,
+    0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,
+    0xfe,0x00,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0xfe,0xff,0xfe,0x00,
+    0xfe,0x00,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0x00,
+    0xfe,0x00,0xfe,0x00,0xfe,0xff,0xfe,0xff,0xfe,0xff,0x00,0xff,0x00,0xff,0x00,0xff,0x00,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0x00,0xfe,0x00,
+    0xfe,0x00,0xfe,0x00,0xfe,0x00,0xfe,0x00,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0xff,0xfe,0x00,0xfe,0x00,0xfe,0x00,0xfe,0x00,
+
 };
 
 const lv_image_dsc_t bluetooth_searching = {
-  .header.cf = LV_COLOR_FORMAT_I1,
-  .header.magic = LV_IMAGE_HEADER_MAGIC,
-  .header.w = 12,
-  .header.h = 17,
-  .data_size = 42,
+  .header = {
+    .magic = LV_IMAGE_HEADER_MAGIC,
+    .cf = LV_COLOR_FORMAT_AL88,
+    .flags = 0,
+    .w = 14,
+    .h = 19,
+    .stride = 28,
+    .reserved_2 = 0,
+  },
+  .data_size = sizeof(bluetooth_searching_map),
   .data = bluetooth_searching_map,
+  .reserved = NULL,
 };
+
