--- boards/shields/nice_view_elemental/src/images/noise_0.c
@@ -1,102 +1,0 @@
-#include <lvgl.h>
-
-#ifndef LV_ATTRIBUTE_MEM_ALIGN
-#define LV_ATTRIBUTE_MEM_ALIGN
-#endif
-
-#ifndef LV_ATTRIBUTE_NOISE_0
-#define LV_ATTRIBUTE_NOISE_0
-#endif
-
-static const
-LV_ATTRIBUTE_MEM_ALIGN LV_ATTRIBUTE_LARGE_CONST LV_ATTRIBUTE_NOISE_0
-uint8_t noise_0_map[] = {
-
-    0xff,0xff,0xff,0xff,0x00,0x00,0x00,0xff,
-
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x42,0x10,0x80,0x82,0x20,0x84,0x21,0x00,0x20,0x81,0x01,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x02,0x48,0x08,0x00,0x08,0x10,0x04,0x00,0x04,0x24,0x80,0x04,0x10,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x20,0x00,0x80,0x00,0x00,0x00,0x00,0x10,0x00,0x00,0x04,0x00,0x00,0x11,0x00,0x00,0x00,
-    0x00,0x00,0x02,0x00,0x02,0x00,0x49,0x02,0x41,0x01,0x00,0xa0,0x80,0x00,0x48,0x01,0x00,0x10,0x00,0x00,
-    0x00,0x00,0x20,0x04,0x00,0x09,0x00,0x10,0x08,0x08,0x08,0x04,0x09,0x20,0x00,0x24,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x10,0x20,0x20,0x00,0x02,0x20,0x40,0x21,0x00,0x20,0x04,0x20,0x80,0x24,0x00,0x40,0x00,
-    0x00,0x01,0x04,0x81,0x09,0x00,0x49,0x40,0x02,0x04,0x80,0x21,0x00,0x21,0x04,0x00,0x80,0x44,0x00,0x00,
-    0x00,0x04,0x00,0x00,0x40,0x12,0x10,0x04,0x48,0xa2,0x09,0x08,0x4a,0x80,0x21,0x21,0x04,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x09,0x08,0x80,0x44,0x81,0x01,0x08,0x04,0x00,0x00,0x14,0x48,0x04,0x10,0x10,0x08,0x00,
-    0x00,0x00,0x24,0x80,0x02,0x55,0x08,0x20,0x84,0x21,0x50,0x92,0xa0,0x41,0x02,0x90,0x01,0x00,0x00,0x40,
-    0x00,0x20,0x00,0x08,0x90,0x00,0x41,0x05,0x28,0x84,0x04,0x48,0x0a,0x24,0x48,0x01,0x44,0x01,0x00,0x00,
-    0x00,0x02,0x02,0x40,0x24,0x92,0x14,0x48,0x42,0x10,0x29,0x04,0x50,0x91,0x00,0xa8,0x00,0x48,0x08,0x00,
-    0x00,0x00,0x20,0x12,0x88,0x21,0x22,0x02,0x94,0xa2,0x40,0x51,0x01,0x08,0x52,0x00,0xa2,0x00,0x40,0x00,
-    0x00,0x20,0x88,0x88,0x22,0x89,0x48,0x40,0x40,0x44,0x24,0x04,0x94,0xa9,0x29,0x04,0x08,0x21,0x00,0x00,
-    0x01,0x02,0x02,0x21,0x48,0x44,0x25,0x15,0x12,0x8a,0x92,0x92,0x01,0x54,0x40,0x68,0x42,0x84,0x00,0x00,
-    0x00,0x00,0x00,0x94,0x84,0x15,0x10,0xc4,0x44,0xa2,0x49,0x48,0xa4,0x02,0x92,0x12,0xa0,0x00,0x04,0x20,
-    0x00,0x08,0xa4,0x02,0x33,0x50,0xc0,0x11,0x48,0xa9,0x29,0x25,0x55,0x55,0x05,0x40,0x08,0x90,0x20,0x00,
-    0x00,0x40,0x01,0x24,0x84,0xa5,0x16,0xaa,0x21,0x10,0x88,0x92,0x29,0x24,0x50,0x8a,0x92,0x02,0x00,0x00,
-    0x02,0x01,0x08,0x82,0x22,0x29,0x20,0x95,0x94,0xc6,0x56,0xa1,0x49,0x52,0xaa,0xa0,0x48,0x90,0x40,0x00,
-    0x00,0x00,0x20,0x29,0x59,0x44,0x54,0x84,0xaa,0x59,0x51,0x14,0x84,0x8a,0xa0,0x2d,0x22,0x21,0x02,0x00,
-    0x00,0x04,0x0a,0x02,0x94,0x2d,0x09,0x22,0x51,0x42,0x55,0x57,0x54,0xea,0x56,0x88,0x20,0x00,0x00,0x48,
-    0x00,0x10,0x81,0x58,0x52,0xa2,0xaa,0xdb,0x15,0x35,0x58,0xd0,0xb4,0x89,0x42,0x85,0x56,0x89,0x10,0x00,
-    0x02,0x40,0x10,0x24,0x24,0x40,0xad,0x28,0x68,0x91,0x14,0x9a,0x05,0x32,0xac,0x54,0x41,0x00,0x40,0x00,
-    0x00,0x04,0x8a,0x4a,0xc5,0x56,0xa1,0x59,0x16,0x65,0x2a,0x4d,0x7c,0x44,0x52,0xab,0x28,0xa9,0x02,0x00,
-    0x00,0x00,0x21,0x29,0x32,0xa9,0x5a,0x20,0xe8,0x99,0x52,0xaa,0x85,0x6a,0xad,0x48,0xa5,0x00,0x20,0x00,
-    0x00,0x11,0x48,0x91,0x59,0x4a,0x2a,0x9a,0x85,0x6e,0xa9,0x55,0x75,0x49,0x71,0x55,0x48,0x54,0x00,0x40,
-    0x02,0x04,0x02,0x8c,0x49,0x31,0xea,0xac,0xb3,0x31,0x6a,0xa2,0x95,0x56,0xca,0xaa,0x25,0x21,0x02,0x00,
-    0x00,0x20,0x81,0x52,0xa9,0x85,0x14,0xa6,0x90,0xc4,0xb6,0x58,0x60,0xaa,0x32,0xd2,0x42,0x48,0x50,0x04,
-    0x00,0x02,0x14,0x92,0x54,0x35,0x52,0x89,0xcf,0x24,0xa1,0xca,0x9a,0x25,0x4a,0x15,0x18,0x10,0x00,0x00,
-    0x02,0x10,0x42,0x50,0x2d,0xc8,0xaa,0x4a,0x99,0xb3,0x5a,0x36,0x63,0xbc,0xa5,0xd4,0xa4,0x85,0x40,0x10,
-    0x00,0x42,0x11,0x0d,0x92,0x13,0x46,0x91,0x48,0x1a,0xa4,0xaa,0x55,0x08,0xaa,0x43,0x52,0xb0,0x08,0x80,
-    0x00,0x00,0x48,0xa0,0xcb,0x54,0xa4,0x66,0x2a,0xe9,0x55,0x2a,0xa9,0xd4,0xb5,0x2a,0x14,0x85,0x20,0x00,
-    0x04,0x08,0x04,0x56,0x20,0x6a,0xad,0xa3,0x85,0x14,0xd0,0xdb,0x55,0x55,0x16,0x91,0x4a,0x10,0x02,0x00,
-    0x00,0x40,0xa5,0x01,0xae,0x94,0xaa,0x48,0x62,0xd6,0x27,0x6a,0x85,0x52,0xc2,0x9a,0xa9,0x48,0x40,0x00,
-    0x00,0x02,0x00,0xd4,0x52,0x9a,0x45,0xb6,0xb9,0x29,0xa8,0xaa,0x55,0xab,0x29,0x50,0x94,0x02,0x00,0x10,
-    0x02,0x10,0x2a,0x13,0x95,0x4b,0xb0,0x53,0x4a,0x95,0x36,0xb1,0x14,0x4d,0x29,0x26,0xa0,0xa8,0x90,0x40,
-    0x08,0x04,0x80,0xaa,0x44,0x92,0x4b,0xa8,0x89,0x49,0x92,0x4a,0xa9,0x61,0x52,0x49,0x26,0x21,0x02,0x00,
-    0x00,0x20,0x14,0x01,0x14,0x55,0xb4,0x2a,0xa1,0x26,0x28,0xb0,0x22,0xaa,0x8a,0x61,0x49,0x40,0x10,0x00,
-    0x00,0x02,0x42,0x54,0x85,0x54,0x15,0x4a,0x9a,0xa5,0x2a,0x8a,0x19,0x15,0x6a,0x2a,0x22,0x10,0x80,0x00,
-    0x00,0x80,0x10,0xa2,0x65,0x4a,0xca,0x28,0x65,0x25,0x49,0x23,0x51,0x52,0x52,0x95,0x14,0x24,0x01,0x00,
-    0x00,0x12,0x89,0x14,0x95,0x29,0x04,0xa2,0xa5,0x54,0x29,0x8c,0x2a,0xa9,0x95,0x44,0x81,0x00,0x10,0x00,
-    0x00,0x00,0x04,0x82,0x22,0x89,0x70,0x12,0x51,0x22,0xaa,0x6a,0xc4,0x48,0x0a,0x89,0x48,0x92,0x40,0x00,
-    0x00,0x00,0x20,0x29,0x28,0x15,0x08,0x40,0x94,0x92,0x55,0x01,0x21,0xa9,0x54,0x25,0x20,0x00,0x00,0x00,
-    0x01,0x09,0x01,0x04,0x05,0x05,0x55,0x56,0x2a,0x49,0x00,0x58,0x48,0x14,0x41,0x28,0x55,0x24,0x08,0x88,
-    0x00,0x00,0x14,0x50,0xa2,0xaa,0x09,0x10,0x91,0x42,0xa6,0x82,0x24,0xa1,0x29,0x40,0x00,0x10,0x80,0x00,
-    0x00,0x00,0x41,0x04,0x08,0x91,0x50,0x85,0x4a,0x12,0x00,0x4a,0x89,0x12,0x44,0x25,0x54,0x00,0x00,0x00,
-    0x00,0x01,0x00,0x41,0x44,0x22,0x24,0x12,0x42,0x44,0x91,0x28,0x20,0x44,0x29,0x20,0x00,0x48,0x10,0x00,
-    0x02,0x10,0x02,0x14,0x42,0x14,0x81,0x55,0x25,0x10,0x44,0x02,0x8a,0x11,0x44,0x92,0xa2,0x00,0x80,0x00,
-    0x00,0x00,0x20,0x41,0x10,0xa0,0x14,0x08,0x00,0x22,0x20,0x90,0x22,0x28,0x21,0x00,0x00,0x00,0x00,0x80,
-    0x00,0x00,0x82,0x00,0x44,0x05,0x42,0x82,0x50,0x88,0x8a,0x22,0x40,0x84,0x80,0x4a,0x24,0x22,0x00,0x00,
-    0x00,0x02,0x00,0x0a,0x01,0x40,0x08,0x21,0x04,0x00,0x40,0x84,0x8a,0x02,0x09,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x08,0x08,0x80,0x48,0x14,0x41,0x40,0x11,0x22,0x11,0x10,0x00,0x90,0x20,0x24,0x91,0x20,0x10,0x00,
-    0x00,0x40,0x00,0x20,0x01,0x00,0x00,0x11,0x00,0x80,0x84,0x01,0x24,0x00,0x81,0x00,0x00,0x00,0x40,0x00,
-    0x00,0x00,0x00,0x02,0x00,0x44,0x52,0x40,0x48,0x10,0x00,0x84,0x00,0x20,0x04,0x22,0x02,0x00,0x00,0x00,
-    0x00,0x00,0x11,0x08,0x24,0x01,0x00,0x08,0x01,0x02,0x48,0x10,0x42,0x88,0x90,0x80,0x48,0x02,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x10,0x09,0x21,0x04,0x08,0x00,0x81,0x08,0x00,0x00,0x08,0x00,0x48,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x89,0x00,0x20,0x00,0x40,0x00,0x22,0x00,0x00,0x04,0x40,0x00,0x01,0x00,0x00,0x00,
-    0x00,0x00,0x80,0x02,0x00,0x04,0x00,0x04,0x00,0x20,0x80,0x04,0x00,0x90,0x04,0x20,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x04,0x40,0x00,0x00,0x80,0x00,0x10,0x02,0x00,0x20,0x22,0x00,0x10,0x00,0x20,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x10,0x40,0x04,0x80,0x01,0x00,0x00,0x80,0x80,0x00,0x00,0x80,0x80,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x08,0x40,0x08,0x02,0x00,0x00,0x00,0x80,0x02,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x20,0x00,0x00,0x20,0x08,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x02,0x01,0x04,0x00,0x00,0x00,0x00,0x20,0x00,0x00,0x04,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x02,0x00,0x00,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x02,0x00,0x40,0x08,0x10,0x00,0x00,0x00,0x20,
-
-};
-
-const lv_image_dsc_t noise_0 = {
-  .header = {
-    .magic = LV_IMAGE_HEADER_MAGIC,
-    .cf = LV_COLOR_FORMAT_I1,
-    .flags = 0,
-    .w = 160,
-    .h = 68,
-    .stride = 20,
-    .reserved_2 = 0,
-  },
-  .data_size = sizeof(noise_0_map),
-  .data = noise_0_map,
-  .reserved = NULL,
-};
-
