--- boards/shields/nice_view_elemental/src/images/noise_1.c
@@ -1,102 +1,0 @@
-#include <lvgl.h>
-
-#ifndef LV_ATTRIBUTE_MEM_ALIGN
-#define LV_ATTRIBUTE_MEM_ALIGN
-#endif
-
-#ifndef LV_ATTRIBUTE_NOISE_1
-#define LV_ATTRIBUTE_NOISE_1
-#endif
-
-static const
-LV_ATTRIBUTE_MEM_ALIGN LV_ATTRIBUTE_LARGE_CONST LV_ATTRIBUTE_NOISE_1
-uint8_t noise_1_map[] = {
-
-    0xff,0xff,0xff,0xff,0x00,0x00,0x00,0xff,
-
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x10,0x44,0x01,0x10,0x02,0x02,0x20,0x84,0x10,0x10,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x01,0x01,0x00,0x48,0x00,0x88,0x20,0x00,0x00,0x42,0x00,0x92,0x10,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x84,0x90,0x00,0x00,0x00,0x20,0x00,0x00,0x84,0x20,0x00,0x22,0x00,0x42,0x20,0x00,0x00,
-    0x00,0x00,0x08,0x00,0x00,0x48,0x08,0x80,0x82,0x20,0x48,0x00,0x84,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x20,0x00,0x02,0x02,0x42,0x0a,0x08,0x80,0x00,0x20,0x00,0x42,0x84,0x22,0x00,0x02,0x00,0x00,
-    0x00,0x00,0x00,0x48,0x80,0x20,0x00,0x00,0x00,0x0a,0x22,0x04,0x01,0x08,0x10,0x80,0x91,0x00,0x00,0x00,
-    0x00,0x01,0x02,0x00,0x09,0x00,0x90,0x48,0x51,0x00,0x80,0x90,0x90,0x00,0x40,0x02,0x00,0x08,0x00,0x00,
-    0x00,0x00,0x08,0x02,0x20,0x44,0x02,0x01,0x04,0x42,0x08,0x00,0x04,0x82,0x12,0x44,0x04,0x00,0x20,0x00,
-    0x00,0x04,0x00,0x00,0x08,0x10,0x48,0x90,0x20,0x88,0x22,0x25,0x41,0x24,0x80,0x20,0x90,0x20,0x02,0x00,
-    0x00,0x00,0x00,0x49,0x42,0x41,0x10,0x05,0x09,0x21,0x40,0x80,0x14,0x02,0x11,0x0a,0x00,0x81,0x00,0x00,
-    0x00,0x10,0x49,0x02,0x08,0x04,0x05,0x20,0x54,0x04,0x12,0x15,0x01,0x48,0x88,0x80,0x40,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x10,0x41,0x12,0x51,0x15,0x01,0x51,0x24,0x40,0x54,0x00,0x22,0x50,0x04,0x24,0x08,0x00,
-    0x00,0x80,0x04,0x04,0x14,0x85,0x12,0x49,0x29,0x04,0x91,0x12,0x82,0xa9,0x10,0x81,0x11,0x00,0x00,0x10,
-    0x00,0x02,0x40,0x91,0x20,0x52,0x44,0xa2,0x44,0x20,0x04,0xa4,0x50,0x44,0x44,0x44,0x40,0x24,0x40,0x00,
-    0x00,0x08,0x11,0x24,0x85,0x09,0x51,0x14,0xa9,0x15,0x52,0x88,0x8e,0x09,0x11,0x21,0x14,0x00,0x00,0x40,
-    0x00,0x20,0x04,0x00,0x51,0x55,0x08,0xaa,0x21,0x52,0x05,0x55,0x21,0x14,0xa4,0x44,0x41,0x54,0x09,0x00,
-    0x00,0x00,0x81,0x55,0x0a,0x42,0xaa,0x08,0x95,0x04,0x92,0x09,0x2c,0x94,0x15,0x11,0x24,0x00,0x80,0x00,
-    0x02,0x04,0x14,0x02,0xa8,0x52,0x85,0x6a,0x84,0x54,0x48,0xa4,0x41,0x49,0x48,0xca,0x90,0x80,0x20,0x00,
-    0x00,0x10,0x40,0xa4,0x01,0x48,0x14,0x8a,0x4a,0x12,0x15,0x41,0x29,0x26,0x0a,0x14,0x04,0x29,0x00,0x00,
-    0x00,0x40,0x08,0x81,0x54,0x2b,0x85,0x51,0x6a,0x65,0x69,0x2d,0x54,0x91,0x95,0x01,0x6a,0x80,0x00,0x00,
-    0x00,0x00,0x22,0x2a,0x26,0x88,0x9a,0x2e,0x95,0x44,0x92,0xa0,0x94,0x98,0x22,0x64,0x90,0x12,0x11,0x10,
-    0x04,0x04,0x88,0x92,0x24,0xa4,0xaa,0x55,0x4a,0x98,0x55,0x54,0x85,0x4c,0xa9,0x2a,0x4a,0x40,0x40,0x00,
-    0x00,0x10,0x02,0x4a,0xa9,0x13,0x29,0x88,0xa8,0x96,0x92,0x93,0x50,0xa1,0x55,0x52,0x20,0x94,0x00,0x00,
-    0x00,0x42,0x49,0x24,0x26,0xaa,0x32,0x55,0x5e,0x61,0x48,0xac,0x49,0x2a,0xaa,0x29,0x96,0x01,0x04,0x40,
-    0x01,0x00,0x82,0x52,0x90,0x95,0x55,0x50,0xa9,0x9a,0x8e,0x69,0x2a,0x96,0xc4,0x8a,0x48,0xa0,0x10,0x00,
-    0x00,0x08,0x24,0x0a,0x4a,0x92,0x54,0xcf,0x54,0x6b,0x55,0x4a,0xca,0xcb,0x32,0x69,0x24,0x82,0x80,0x00,
-    0x04,0x01,0x02,0x61,0x49,0x53,0x13,0x30,0xa1,0xa2,0xa8,0xa4,0xa4,0xb4,0x89,0x41,0x44,0x50,0x12,0x10,
-    0x00,0x20,0x14,0x0a,0x69,0x59,0x56,0xa5,0x3e,0x18,0x96,0xa2,0x2a,0x88,0xac,0x24,0x23,0x4a,0x40,0x40,
-    0x00,0x09,0x41,0x92,0x89,0x22,0xaa,0xaa,0x48,0xb2,0x52,0x9a,0xe6,0xcb,0x33,0x43,0x90,0x20,0x00,0x00,
-    0x04,0x00,0x14,0x49,0xb6,0xc8,0xad,0x1d,0x0a,0x5c,0x95,0x48,0x10,0x2c,0xc8,0xbc,0x2a,0x85,0x22,0x00,
-    0x00,0x24,0x82,0xaa,0x81,0x33,0x32,0xd5,0x4c,0xaa,0xaa,0x69,0x6b,0x95,0x29,0x22,0x80,0x90,0x08,0x00,
-    0x00,0x80,0x24,0x44,0x7b,0x8c,0xcd,0x52,0x73,0x53,0x55,0x82,0x94,0xaa,0xaa,0xd4,0xad,0x02,0x80,0x00,
-    0x00,0x00,0x80,0x36,0x2c,0x62,0xa2,0x6d,0x48,0x29,0x6d,0x2c,0x43,0x95,0x55,0x52,0x10,0x50,0x00,0x40,
-    0x02,0x0a,0x15,0x02,0xc9,0x84,0xcb,0xb0,0xa8,0xea,0xa9,0x2b,0x5a,0xa9,0x95,0x45,0xaa,0x00,0xa4,0x00,
-    0x08,0x20,0x40,0xa8,0x2a,0x2b,0x48,0x96,0x57,0x0a,0xc8,0xa2,0x69,0x12,0xdb,0x32,0x42,0xa4,0x00,0x00,
-    0x00,0x01,0x12,0x41,0xa2,0xb4,0xb6,0xc9,0x21,0x21,0x6a,0x9b,0x95,0xa5,0x04,0xc8,0x92,0x48,0x80,0x04,
-    0x00,0x08,0x00,0x14,0x88,0x25,0x46,0x6c,0xa4,0x5a,0x90,0xa4,0xa4,0xd1,0x72,0x16,0x48,0x84,0x10,0x10,
-    0x00,0x80,0xa9,0x41,0x25,0x54,0xa9,0x21,0x31,0xa5,0x6d,0x32,0x4a,0x28,0x8a,0xa1,0x12,0x28,0x82,0x00,
-    0x00,0x00,0x02,0x25,0x54,0x8a,0xaa,0xcd,0x4d,0x2d,0x45,0x28,0xaa,0xc7,0x48,0xb4,0x64,0x92,0x10,0x00,
-    0x04,0x09,0x10,0x22,0x05,0x48,0x54,0xa8,0xa0,0x52,0x92,0x68,0x92,0x28,0x26,0x41,0x00,0x40,0x00,0x00,
-    0x00,0x40,0x05,0x11,0xb0,0x35,0x12,0x92,0x5a,0xa5,0x49,0x2b,0x55,0xa1,0x29,0x2a,0x55,0x14,0x80,0x20,
-    0x00,0x00,0x48,0x48,0x56,0x22,0x55,0x42,0x84,0x90,0xa8,0x04,0x40,0x85,0x54,0x88,0x84,0x00,0x00,0x80,
-    0x00,0x04,0x02,0x09,0x01,0x50,0x94,0x52,0xa8,0x24,0x25,0x4a,0xa0,0xa9,0x42,0xa6,0x51,0x28,0x22,0x00,
-    0x00,0x80,0x91,0x40,0x54,0x89,0x29,0x48,0x12,0xa5,0x50,0x24,0x36,0x44,0x38,0x02,0x08,0x01,0x00,0x00,
-    0x00,0x00,0x02,0x15,0x22,0x22,0x20,0x92,0xa4,0x85,0x0b,0x42,0x89,0x35,0x05,0x51,0x40,0x84,0x00,0x00,
-    0x00,0x10,0x20,0x40,0x94,0x90,0x95,0x28,0x09,0x50,0xa4,0x14,0x24,0x48,0xa0,0x2a,0x24,0x00,0x20,0x00,
-    0x02,0x00,0x81,0x09,0x40,0x24,0x42,0x02,0x90,0x05,0x51,0x02,0x90,0x01,0x49,0x01,0x11,0x21,0x00,0x00,
-    0x00,0x02,0x04,0x04,0x14,0x85,0x08,0x94,0x24,0xa8,0x02,0x49,0x25,0x24,0x92,0x20,0x40,0x00,0x04,0x00,
-    0x00,0x08,0x10,0x91,0x02,0x10,0x24,0x41,0x02,0x05,0x10,0x80,0x08,0x11,0x00,0x94,0x10,0x22,0x00,0x20,
-    0x00,0x00,0x00,0x04,0x50,0x44,0x10,0x08,0x50,0x90,0x25,0x12,0xa1,0x44,0x55,0x01,0x42,0x00,0x00,0x00,
-    0x00,0x00,0x41,0x01,0x04,0x81,0x42,0x41,0x04,0x04,0x80,0x00,0x00,0x01,0x00,0x48,0x00,0x40,0x20,0x00,
-    0x00,0x40,0x04,0x20,0x00,0x12,0x08,0x10,0x00,0x90,0x02,0x52,0x10,0x90,0x29,0x02,0x10,0x08,0x00,0x00,
-    0x00,0x00,0x00,0x04,0x50,0x40,0x81,0x44,0x42,0x05,0x28,0x00,0x84,0x00,0x80,0x40,0x42,0x01,0x00,0x00,
-    0x00,0x00,0x80,0x00,0x02,0x04,0x24,0x01,0x10,0x40,0x01,0x02,0x01,0x22,0x04,0x08,0x00,0x00,0x00,0x00,
-    0x00,0x02,0x02,0x41,0x10,0x90,0x80,0x10,0x00,0x09,0x20,0x20,0x48,0x00,0x20,0x40,0x88,0x40,0x02,0x00,
-    0x00,0x00,0x00,0x04,0x00,0x00,0x04,0x00,0x41,0x20,0x04,0x00,0x00,0x88,0x01,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x08,0x00,0x04,0x02,0x10,0x89,0x08,0x00,0x00,0x88,0x90,0x01,0x04,0x04,0x00,0x04,0x00,0x00,
-    0x00,0x00,0x00,0x10,0x20,0x40,0x00,0x00,0x00,0x80,0x40,0x02,0x02,0x20,0x10,0x10,0x12,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x04,0x40,0x00,0x00,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x40,0x00,0x10,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x80,0x44,0x08,0x04,0x20,0x00,0x04,0x00,0x40,0x00,0x10,0x00,0x00,
-    0x00,0x10,0x00,0x41,0x02,0x20,0x04,0x08,0x00,0x00,0x00,0x00,0x44,0x10,0x40,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x10,0x00,0x00,0x40,0x00,0x02,0x00,0x40,0x02,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x00,0x00,0x08,0x80,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x00,
-    0x00,0x00,0x10,0x00,0x08,0x00,0x00,0x00,0x80,0x00,0x20,0x00,0x00,0x00,0x00,0x01,0x08,0x00,0x00,0x00,
-
-};
-
-const lv_image_dsc_t noise_1 = {
-  .header = {
-    .magic = LV_IMAGE_HEADER_MAGIC,
-    .cf = LV_COLOR_FORMAT_I1,
-    .flags = 0,
-    .w = 160,
-    .h = 68,
-    .stride = 20,
-    .reserved_2 = 0,
-  },
-  .data_size = sizeof(noise_1_map),
-  .data = noise_1_map,
-  .reserved = NULL,
-};
-
