--- boards/shields/nice_view_elemental/src/images/noise_2.c
@@ -1,102 +1,0 @@
-#include <lvgl.h>
-
-#ifndef LV_ATTRIBUTE_MEM_ALIGN
-#define LV_ATTRIBUTE_MEM_ALIGN
-#endif
-
-#ifndef LV_ATTRIBUTE_NOISE_2
-#define LV_ATTRIBUTE_NOISE_2
-#endif
-
-static const
-LV_ATTRIBUTE_MEM_ALIGN LV_ATTRIBUTE_LARGE_CONST LV_ATTRIBUTE_NOISE_2
-uint8_t noise_2_map[] = {
-
-    0xff,0xff,0xff,0xff,0x00,0x00,0x00,0xff,
-
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x10,0x88,0x90,0x88,0x01,0x08,0x21,0x10,0x84,0x10,0x20,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x24,0x80,0x00,0x00,0x00,0x08,0x00,0x00,0x02,0x00,0x40,0x80,0x80,0x80,0x00,0x00,
-    0x00,0x00,0x00,0x88,0x80,0x02,0x00,0x02,0x00,0x20,0x20,0x84,0x00,0x21,0x02,0x04,0x08,0x00,0x00,0x00,
-    0x00,0x00,0x04,0x00,0x00,0x00,0x49,0x20,0x4a,0x82,0x02,0x00,0x00,0x04,0x00,0x00,0x20,0x00,0x00,0x00,
-    0x00,0x00,0x10,0x02,0x00,0x10,0x00,0x00,0x00,0x00,0x40,0x10,0x88,0x80,0x28,0x10,0x00,0x04,0x00,0x00,
-    0x00,0x00,0x40,0x08,0x44,0x82,0x04,0x22,0x00,0x12,0x04,0x00,0x22,0x00,0x80,0x82,0x42,0x10,0x40,0x00,
-    0x00,0x00,0x00,0x40,0x00,0x20,0x41,0x00,0x94,0x80,0x20,0x44,0x00,0x24,0x02,0x04,0x08,0x00,0x00,0x00,
-    0x00,0x02,0x01,0x01,0x11,0x04,0x88,0x24,0x00,0x49,0x09,0x01,0x10,0x00,0x28,0x20,0x80,0x00,0x00,0x00,
-    0x00,0x00,0x04,0x14,0x44,0x00,0x20,0x01,0x11,0x20,0x40,0x28,0x44,0x42,0x80,0x82,0x12,0x10,0x08,0x00,
-    0x00,0x10,0x20,0x00,0x10,0x91,0x05,0x20,0x20,0x02,0x15,0x02,0x09,0x28,0x10,0x28,0x40,0x40,0x80,0x40,
-    0x00,0x00,0x00,0x84,0x80,0x20,0xa0,0x0a,0x04,0x90,0x00,0x48,0x80,0x01,0x04,0x01,0x08,0x04,0x00,0x00,
-    0x00,0x01,0x02,0x20,0x15,0x44,0x12,0x88,0xa0,0x25,0x50,0x01,0x2a,0xa0,0xa2,0xa4,0x21,0x00,0x00,0x00,
-    0x00,0x04,0x20,0x04,0x80,0x02,0xa8,0x22,0x2a,0x80,0x0a,0x94,0x45,0x0a,0x11,0x09,0x00,0x22,0x00,0x00,
-    0x00,0x40,0x09,0x40,0x51,0x54,0x05,0x11,0x08,0x4a,0xa1,0x0a,0x90,0x40,0xa4,0x20,0x29,0x00,0x24,0x00,
-    0x00,0x00,0x00,0x12,0x84,0x84,0xa8,0x85,0x49,0x09,0x2a,0x10,0x48,0xaa,0x51,0x4a,0x80,0x28,0x00,0x00,
-    0x00,0x00,0x22,0x88,0x25,0x52,0x92,0x42,0x20,0x54,0x80,0x55,0x45,0x20,0x8a,0x0a,0x15,0x01,0x00,0x20,
-    0x00,0x81,0x08,0x25,0x12,0x50,0x41,0x29,0x55,0x02,0x65,0x24,0x14,0x2c,0xa9,0xd0,0x80,0x50,0x20,0x00,
-    0x02,0x08,0x01,0x01,0x41,0x2a,0xaa,0xa4,0x88,0xa9,0x15,0x11,0x91,0x21,0x2a,0x12,0x4a,0x01,0x02,0x00,
-    0x00,0x00,0x42,0x54,0x4a,0x88,0x52,0x2a,0xa3,0x24,0xaa,0xa6,0x52,0x5a,0x94,0x89,0x20,0x94,0x08,0x00,
-    0x00,0x12,0x11,0x12,0x49,0x52,0xa4,0x15,0x2a,0x54,0x25,0x50,0x05,0x12,0x92,0x22,0x0a,0x40,0x80,0x00,
-    0x00,0x40,0x02,0x48,0x64,0x81,0x2a,0xc8,0x85,0xa9,0xa8,0x8a,0xe8,0xda,0xa8,0x94,0x48,0x12,0x20,0x20,
-    0x01,0x01,0x48,0x14,0x92,0x95,0x48,0x24,0x69,0x4d,0x56,0x51,0x4a,0xa2,0x4a,0x52,0x85,0x40,0x04,0x00,
-    0x00,0x04,0x22,0xa0,0xa8,0xd0,0xa6,0x95,0x55,0x52,0x89,0x96,0x95,0x2b,0x2a,0x20,0x24,0x94,0x80,0x00,
-    0x00,0x00,0x00,0x15,0x12,0x4a,0xa0,0x72,0xaa,0x94,0x24,0xa5,0x95,0x94,0xa5,0x1a,0xa1,0x10,0x10,0x80,
-    0x04,0x44,0x94,0xa2,0x48,0xa4,0x97,0x09,0x2c,0xa4,0xa5,0x5a,0x52,0xca,0x12,0x68,0x2a,0x42,0x00,0x00,
-    0x00,0x00,0x20,0x98,0xa7,0x32,0x54,0xc5,0x4a,0x95,0xd5,0x65,0x34,0xa4,0xd5,0x4a,0x84,0x11,0x02,0x00,
-    0x00,0x08,0x89,0x05,0x35,0x58,0x65,0x36,0xb1,0x2a,0x14,0x25,0x4a,0x97,0x2a,0xa8,0xb5,0x48,0x20,0x00,
-    0x00,0x82,0x21,0x52,0x49,0x2b,0xa4,0xe9,0x56,0xc9,0xb5,0xcc,0x89,0x28,0xab,0x4a,0x08,0x82,0x08,0x08,
-    0x02,0x08,0x02,0x29,0x24,0xa8,0x57,0x56,0x55,0x61,0x49,0x53,0xaa,0x9a,0x90,0x81,0xa2,0x20,0x80,0x40,
-    0x00,0x00,0x24,0x8a,0x54,0xa2,0x48,0x4b,0x25,0x52,0x6d,0x34,0x45,0x45,0xe4,0x2c,0x14,0x15,0x20,0x00,
-    0x00,0x21,0x10,0x21,0x2a,0xdd,0x52,0xa5,0x55,0xad,0x4a,0x85,0x74,0xac,0x0d,0xeb,0xa1,0x40,0x00,0x00,
-    0x08,0x80,0x42,0xd2,0x8e,0x44,0xcd,0x15,0x29,0x11,0x77,0x7b,0x97,0x2a,0x95,0x24,0x8a,0x92,0x12,0x00,
-    0x00,0x09,0x12,0x41,0x60,0xbb,0x6a,0xb4,0xb5,0x57,0x42,0x44,0x54,0xc5,0x66,0xa9,0x55,0x20,0x80,0x10,
-    0x00,0x00,0x49,0x15,0x15,0x04,0x54,0x82,0x54,0x50,0xd2,0xb2,0x59,0x57,0x59,0x49,0x20,0x94,0x00,0x40,
-    0x00,0x85,0x02,0x44,0xaa,0xd3,0x17,0x5d,0x6b,0x4e,0x3a,0x55,0x4a,0x99,0x2a,0x46,0x12,0x01,0x09,0x00,
-    0x02,0x00,0x09,0x2a,0x55,0x32,0xa9,0x2a,0x14,0xa0,0x84,0x52,0x7a,0x94,0xd4,0xb5,0x54,0x54,0x20,0x00,
-    0x00,0x11,0x42,0x92,0x92,0xca,0x6d,0x52,0xca,0x2c,0x28,0x41,0x0d,0xa5,0x0a,0x28,0xa9,0x00,0x80,0x00,
-    0x00,0x00,0x10,0x48,0x4a,0x22,0x92,0x89,0x51,0xc5,0x95,0x2a,0xd2,0x54,0x45,0x40,0x20,0x4a,0x00,0x00,
-    0x00,0x41,0x0a,0x89,0x2a,0xcd,0x52,0xac,0xaa,0x08,0xaa,0x55,0x29,0x4a,0x93,0x2a,0x95,0x20,0x21,0x00,
-    0x04,0x04,0x20,0x22,0x48,0x48,0x95,0x53,0x25,0x35,0x22,0xb2,0x0c,0x95,0x28,0x89,0x10,0x11,0x00,0x00,
-    0x00,0x00,0x05,0x04,0xaa,0xa6,0x4a,0x54,0x89,0xa2,0xaa,0xa9,0x50,0x68,0xd5,0x54,0xaa,0x40,0x08,0x04,
-    0x00,0x22,0x50,0x52,0x24,0x91,0x2a,0x92,0xa0,0x88,0x14,0x84,0xa5,0x26,0x00,0x42,0x10,0x12,0x20,0x20,
-    0x00,0x00,0x01,0x01,0x00,0x4c,0x90,0x44,0x16,0x17,0x26,0x32,0x84,0x81,0x51,0xa0,0x85,0x40,0x00,0x80,
-    0x00,0x01,0x08,0x2c,0x55,0x51,0x0d,0x32,0x81,0x40,0x08,0x88,0x91,0x54,0x4a,0x0a,0x20,0x02,0x40,0x00,
-    0x01,0x10,0x02,0x01,0x10,0x80,0x92,0x48,0x30,0xa9,0xa2,0x64,0x28,0x8b,0x29,0x40,0x8a,0x90,0x02,0x00,
-    0x00,0x00,0x48,0x50,0x42,0x29,0x44,0xa2,0x84,0x44,0x14,0x11,0x05,0x24,0x40,0x2a,0x50,0x02,0x08,0x00,
-    0x00,0x01,0x00,0x04,0x28,0x80,0x12,0x04,0x29,0x14,0xa2,0x92,0x48,0x52,0x8c,0x82,0x02,0x20,0x40,0x00,
-    0x00,0x20,0x00,0x91,0x0a,0x4a,0x84,0x24,0x80,0x85,0x10,0x28,0x20,0x00,0x02,0x28,0x90,0x81,0x00,0x00,
-    0x00,0x80,0x24,0x04,0xa0,0x20,0x48,0x92,0x25,0x20,0x0a,0x01,0x0a,0x94,0xa4,0x82,0x04,0x08,0x00,0x00,
-    0x02,0x02,0x00,0x40,0x04,0x84,0x00,0x05,0x10,0x49,0x40,0x88,0x40,0x01,0x08,0x28,0x20,0x20,0x00,0x00,
-    0x00,0x00,0x00,0x02,0x08,0x11,0x49,0x50,0x41,0x00,0x12,0x21,0x14,0x90,0x81,0x01,0x00,0x80,0x08,0x00,
-    0x00,0x00,0x48,0x90,0xa1,0x44,0x04,0x01,0x04,0x22,0x04,0x04,0x40,0x04,0x24,0x48,0x10,0x00,0x40,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0xa1,0x08,0x20,0x88,0x50,0x90,0x04,0x40,0x01,0x00,0x42,0x02,0x00,0x00,
-    0x00,0x02,0x02,0x04,0x80,0x20,0x08,0x41,0x00,0x00,0x00,0x01,0x01,0x11,0x48,0x04,0x00,0x10,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x11,0x02,0x00,0x10,0x04,0x84,0x88,0x44,0x44,0x20,0x00,0x90,0x08,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x21,0x04,0x20,0x25,0x04,0x20,0x10,0x01,0x10,0x10,0x02,0x20,0x00,0x00,0x00,0x00,0x10,
-    0x00,0x00,0x10,0x80,0x00,0x01,0x00,0x00,0x80,0x00,0x20,0x00,0x01,0x00,0x00,0x01,0x20,0x80,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x08,0x00,0x10,0x02,0x01,0x00,0x01,0x20,0x10,0x04,0x04,0x02,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x40,0x00,0x41,0x00,0x00,0x20,0x04,0x44,0x00,0x00,0x80,0x40,0x00,0x02,0x20,0x80,
-    0x00,0x10,0x81,0x04,0x02,0x00,0x08,0x00,0x88,0x00,0x10,0x00,0x04,0x44,0x08,0x00,0x00,0x08,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x08,0x02,0x00,0x02,0x00,0x04,0x40,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x20,0x00,0x40,0x00,0x80,0x00,0x10,0x00,0x00,0x00,0x10,0x40,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x40,0x00,0x40,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x40,0x00,0x80,0x02,0x00,0x02,0x00,0x00,0x00,
-
-};
-
-const lv_image_dsc_t noise_2 = {
-  .header = {
-    .magic = LV_IMAGE_HEADER_MAGIC,
-    .cf = LV_COLOR_FORMAT_I1,
-    .flags = 0,
-    .w = 160,
-    .h = 68,
-    .stride = 20,
-    .reserved_2 = 0,
-  },
-  .data_size = sizeof(noise_2_map),
-  .data = noise_2_map,
-  .reserved = NULL,
-};
-
