--- boards/shields/nice_view_elemental/src/images/noise_3.c
@@ -1,102 +1,0 @@
-#include <lvgl.h>
-
-#ifndef LV_ATTRIBUTE_MEM_ALIGN
-#define LV_ATTRIBUTE_MEM_ALIGN
-#endif
-
-#ifndef LV_ATTRIBUTE_NOISE_3
-#define LV_ATTRIBUTE_NOISE_3
-#endif
-
-static const
-LV_ATTRIBUTE_MEM_ALIGN LV_ATTRIBUTE_LARGE_CONST LV_ATTRIBUTE_NOISE_3
-uint8_t noise_3_map[] = {
-
-    0xff,0xff,0xff,0xff,0x00,0x00,0x00,0xff,
-
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x10,0x41,0x11,0x10,0x82,0x10,0x40,0x42,0x21,0x04,0x08,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x04,0x40,0x04,0x00,0x00,0x00,0x01,0x09,0x00,0x00,0x10,0x40,0x49,0x08,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x20,0x02,0x00,0x04,0x02,0x08,0x40,0x00,0x08,0x08,0x40,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x01,0x00,0x00,0x48,0x90,0x40,0x20,0x08,0x00,0x01,0x41,0x00,0x01,0x00,0x10,0x88,0x00,0x00,
-    0x00,0x00,0x10,0x01,0x10,0x00,0x00,0x00,0x01,0x00,0x84,0x44,0x00,0x09,0x10,0x08,0x40,0x00,0x40,0x00,
-    0x00,0x00,0x40,0x04,0x42,0x04,0x40,0x92,0x48,0x02,0x11,0x10,0x00,0x80,0x04,0x21,0x02,0x00,0x00,0x00,
-    0x00,0x01,0x00,0x80,0x00,0x81,0x04,0x00,0x00,0x50,0x40,0x00,0xa4,0x21,0x20,0x80,0x00,0x02,0x00,0x00,
-    0x00,0x00,0x02,0x10,0x08,0x10,0x20,0x88,0x48,0x00,0x00,0x24,0x01,0x04,0x08,0x04,0x20,0x48,0x00,0x00,
-    0x00,0x00,0x00,0x04,0x42,0x49,0x12,0x21,0x22,0x8a,0x4a,0x42,0x12,0x00,0x40,0x10,0x09,0x00,0x04,0x00,
-    0x00,0x04,0x10,0x21,0x10,0x00,0x80,0x04,0x80,0x40,0x00,0x08,0x80,0x52,0x8a,0x81,0x20,0x00,0x10,0x00,
-    0x00,0x10,0x02,0x00,0x42,0x50,0x25,0x20,0x12,0x15,0x4a,0x42,0x25,0x08,0x20,0x14,0x04,0x10,0x00,0x00,
-    0x00,0x00,0x20,0x29,0x08,0x0a,0x80,0x92,0x81,0x40,0x20,0x28,0x48,0x22,0x49,0x41,0x11,0x02,0x40,0x00,
-    0x00,0x00,0x80,0x80,0x41,0x41,0x24,0x0a,0x28,0x15,0x89,0x12,0x80,0x48,0x84,0x14,0x40,0x40,0x00,0x08,
-    0x00,0x00,0x0a,0x0a,0x12,0x54,0x12,0x44,0x44,0x48,0x21,0x08,0x25,0x22,0x22,0x80,0x12,0x08,0x02,0x00,
-    0x00,0x44,0x00,0x10,0x42,0x25,0xa1,0x55,0x15,0x24,0x45,0x52,0xa8,0x52,0x10,0x55,0x40,0x80,0x80,0x00,
-    0x00,0x00,0x41,0x04,0x98,0x90,0x95,0x02,0xa0,0x92,0x88,0x90,0x82,0x88,0xaa,0x08,0x08,0x24,0x10,0x20,
-    0x00,0x01,0x08,0x22,0x02,0x46,0x4a,0x95,0x49,0x09,0x28,0x95,0x24,0x54,0x05,0x52,0xa2,0x80,0x40,0x00,
-    0x01,0x04,0x05,0x08,0x89,0x21,0x05,0x48,0x25,0x64,0xc5,0x11,0x15,0x15,0x52,0x08,0x14,0x05,0x00,0x00,
-    0x00,0x10,0x20,0x85,0x2a,0x9a,0xa9,0x2d,0x51,0x14,0x15,0x4c,0xa5,0xc1,0x22,0x91,0x81,0x10,0x04,0x40,
-    0x00,0x00,0x88,0x52,0x80,0x42,0xa4,0x42,0xac,0xa2,0xe4,0x92,0x10,0x2a,0x54,0x46,0x2a,0x00,0x10,0x00,
-    0x04,0x02,0x02,0x08,0x55,0x95,0x52,0xa8,0x91,0x24,0x12,0x0a,0xc9,0xd4,0xab,0x70,0xa4,0xa9,0x00,0x00,
-    0x00,0x40,0x48,0xa2,0xa8,0x46,0xaa,0x45,0x55,0x51,0xa8,0xd1,0x69,0x25,0x44,0x0a,0x50,0x00,0x41,0x00,
-    0x00,0x08,0x81,0x09,0x15,0x50,0x48,0xb4,0x48,0x96,0x27,0x54,0x95,0x35,0x5a,0xeb,0x12,0xa2,0x00,0x00,
-    0x00,0x02,0x24,0xa5,0x4a,0x49,0x0d,0x12,0xa5,0x84,0xb9,0x4b,0x28,0xca,0x85,0x04,0x44,0x14,0x24,0x08,
-    0x02,0x10,0x00,0x11,0x22,0xa5,0x73,0xb1,0x4a,0x7b,0x4d,0x4a,0xc6,0x2a,0x4a,0xa9,0x2a,0x80,0x80,0x00,
-    0x00,0x41,0x12,0xa5,0x5a,0xd3,0x4a,0xa8,0xaa,0x88,0x91,0x32,0xba,0x90,0xd5,0x28,0xc9,0x54,0x00,0x40,
-    0x00,0x00,0x24,0x12,0x25,0x29,0x35,0x4d,0x50,0xb5,0x54,0xd5,0x45,0x55,0x48,0xc4,0x54,0x20,0x88,0x00,
-    0x00,0x04,0x88,0x92,0x95,0x55,0x4a,0xb2,0xae,0x96,0x52,0x96,0xaa,0xa5,0x67,0x22,0x49,0x04,0x20,0x00,
-    0x04,0x80,0x22,0x49,0x56,0x29,0x75,0x49,0x15,0x62,0xa6,0x24,0x55,0xd8,0xb8,0xb5,0x24,0xa9,0x01,0x00,
-    0x00,0x12,0x08,0x91,0x22,0x8b,0x92,0x5c,0x95,0xb8,0x35,0x5a,0x5a,0x2e,0xc8,0x81,0x22,0x42,0x10,0x04,
-    0x00,0x00,0x22,0x29,0x19,0x65,0x41,0x50,0xa8,0xa7,0x49,0x88,0xcb,0xb0,0x73,0x6c,0xa8,0x10,0x02,0x00,
-    0x04,0x22,0x89,0x02,0xad,0x2a,0xac,0x81,0x52,0x91,0x36,0x65,0x3a,0x8b,0x8d,0x13,0x4a,0xa4,0x80,0x00,
-    0x00,0x08,0x24,0x54,0x8a,0x95,0xb6,0xae,0x6e,0x6d,0x4a,0xb6,0x88,0x55,0x55,0xc2,0x81,0x00,0x08,0x20,
-    0x00,0x80,0x82,0x95,0x55,0x56,0x42,0x51,0xa3,0x16,0x1a,0xa2,0xd5,0xa9,0x42,0xb9,0x74,0x2a,0x90,0x00,
-    0x00,0x02,0x14,0x22,0x49,0x29,0x55,0x88,0x28,0xd9,0xd6,0xba,0xaa,0x15,0xb2,0x05,0x0b,0x10,0x00,0x00,
-    0x08,0x20,0x02,0x89,0x55,0x2c,0x92,0x4e,0x52,0xa2,0x49,0x45,0x25,0x2a,0x43,0xaa,0xe4,0x84,0x22,0x00,
-    0x00,0x00,0xa8,0x09,0x2a,0xa2,0xa5,0x55,0x55,0x13,0x54,0x56,0x95,0x51,0x5a,0xa8,0x09,0x11,0x00,0x20,
-    0x00,0x88,0x05,0x45,0xa5,0x52,0x8a,0xa1,0x42,0xd5,0x2a,0x80,0xd6,0x55,0x42,0x83,0x28,0x00,0x48,0x00,
-    0x02,0x01,0x20,0x30,0x12,0x20,0x94,0xd8,0xa9,0x44,0x95,0x2a,0xa1,0x5a,0xbc,0x28,0xa1,0x29,0x00,0x00,
-    0x00,0x04,0x04,0x8b,0xaa,0xb5,0x62,0x0d,0x25,0x64,0xc8,0x54,0x0e,0x8a,0xa1,0x0a,0x0c,0x80,0x01,0x00,
-    0x00,0x10,0x11,0x20,0x51,0x44,0x44,0x25,0x14,0x85,0x32,0xa5,0xb1,0x45,0x15,0x51,0x42,0x20,0x90,0x00,
-    0x00,0x00,0x40,0x15,0x4c,0xb3,0x12,0xa4,0xa2,0x75,0x28,0x91,0x22,0x48,0xc4,0x08,0x15,0x0a,0x00,0x08,
-    0x04,0x42,0x09,0x20,0x22,0x89,0x05,0x15,0x12,0x8a,0x44,0x58,0xaa,0x08,0x2a,0x54,0x80,0x40,0x04,0x00,
-    0x00,0x00,0x20,0x0a,0x08,0x50,0xa5,0x44,0x94,0x40,0xaa,0x82,0x41,0x46,0x91,0x02,0x50,0x01,0x10,0x00,
-    0x00,0x04,0x04,0x90,0xaa,0x8d,0x54,0x2a,0x24,0x56,0x10,0x50,0x94,0xa0,0x4a,0x4a,0x8a,0x94,0x00,0x20,
-    0x00,0x00,0x90,0x08,0x80,0x92,0x02,0x21,0x01,0x09,0x25,0x14,0x22,0x55,0x00,0xa4,0x20,0x00,0x40,0x00,
-    0x00,0x10,0x01,0x44,0x25,0x08,0x90,0x44,0x94,0x50,0x84,0x42,0x89,0x12,0xaa,0x11,0x04,0x44,0x00,0x00,
-    0x00,0x80,0x08,0x11,0x42,0x22,0x42,0x10,0x22,0x05,0x28,0x10,0x40,0x90,0x01,0x40,0x80,0x90,0x00,0x00,
-    0x00,0x00,0x80,0x40,0x08,0x89,0x20,0xa5,0x49,0x50,0x89,0x4a,0x2a,0x45,0x50,0x14,0x22,0x00,0x48,0x80,
-    0x00,0x02,0x08,0x05,0x42,0x00,0x8a,0x00,0x00,0x02,0x00,0x00,0x84,0x00,0x85,0x02,0x80,0x01,0x00,0x00,
-    0x00,0x00,0x00,0x90,0x10,0x55,0x20,0x4a,0x89,0x28,0x52,0x89,0x11,0x52,0x00,0x10,0x04,0x10,0x00,0x00,
-    0x00,0x00,0x10,0x00,0x82,0x00,0x05,0x10,0x20,0x01,0x00,0x20,0x44,0x00,0xa4,0x40,0x81,0x00,0x00,0x00,
-    0x01,0x10,0x40,0x82,0x08,0x8a,0x20,0x02,0x89,0x14,0x44,0x85,0x00,0x44,0x00,0x04,0x10,0x20,0x00,0x00,
-    0x00,0x00,0x00,0x08,0x00,0x00,0x00,0x40,0x00,0x40,0x00,0x00,0x12,0x00,0x82,0x10,0x40,0x04,0x08,0x00,
-    0x00,0x01,0x02,0x00,0x22,0x10,0x89,0x04,0x21,0x00,0x20,0x10,0x80,0x10,0x10,0x42,0x00,0x00,0x80,0x00,
-    0x00,0x00,0x00,0x00,0x80,0x40,0x00,0x10,0x00,0x12,0x04,0x42,0x00,0x42,0x40,0x00,0x04,0x20,0x00,0x00,
-    0x00,0x00,0x00,0x22,0x00,0x02,0x01,0x00,0x88,0x40,0x00,0x00,0x09,0x00,0x02,0x00,0x80,0x00,0x00,0x00,
-    0x00,0x00,0x04,0x00,0x08,0x00,0x44,0x02,0x00,0x00,0x42,0x10,0x20,0x00,0x00,0x10,0x00,0x80,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x84,0x00,0x10,0x00,0x02,0x08,0x00,0x00,0x08,0x90,0x40,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x20,0x01,0x00,0x00,0x00,0x00,0x10,0x80,0x00,0x00,0x00,0x00,0x00,0x01,0x04,0x00,0x00,0x00,
-    0x00,0x04,0x00,0x10,0x10,0x00,0x40,0x00,0x00,0x10,0x00,0x41,0x04,0x20,0x02,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x40,0x00,0x10,0x02,0x02,0x00,0x00,0x02,0x00,0x10,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x40,0x00,0x10,0x02,0x00,0x88,0x04,0x00,0x00,0x20,0x04,0x10,0x01,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x20,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-
-};
-
-const lv_image_dsc_t noise_3 = {
-  .header = {
-    .magic = LV_IMAGE_HEADER_MAGIC,
-    .cf = LV_COLOR_FORMAT_I1,
-    .flags = 0,
-    .w = 160,
-    .h = 68,
-    .stride = 20,
-    .reserved_2 = 0,
-  },
-  .data_size = sizeof(noise_3_map),
-  .data = noise_3_map,
-  .reserved = NULL,
-};
-
