--- boards/shields/nice_view_elemental/src/images/noise_4.c
@@ -1,102 +1,0 @@
-#include <lvgl.h>
-
-#ifndef LV_ATTRIBUTE_MEM_ALIGN
-#define LV_ATTRIBUTE_MEM_ALIGN
-#endif
-
-#ifndef LV_ATTRIBUTE_NOISE_4
-#define LV_ATTRIBUTE_NOISE_4
-#endif
-
-static const
-LV_ATTRIBUTE_MEM_ALIGN LV_ATTRIBUTE_LARGE_CONST LV_ATTRIBUTE_NOISE_4
-uint8_t noise_4_map[] = {
-
-    0xff,0xff,0xff,0xff,0x00,0x00,0x00,0xff,
-
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x01,0x08,0x80,0x21,0x04,0x10,0x84,0x42,0x10,0x04,0x08,0x80,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x11,0x08,0x00,0x10,0x84,0x10,0x40,0x10,0x08,0x01,0x20,0x40,0x12,0x40,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x80,0x20,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x00,0x00,0x00,0x00,0x04,0x00,0x00,0x00,
-    0x00,0x00,0x02,0x00,0x00,0x89,0x00,0x00,0x20,0x80,0x00,0x80,0x48,0x02,0x01,0x00,0x00,0x21,0x00,0x00,
-    0x00,0x00,0x20,0x01,0x00,0x00,0x20,0x24,0x00,0x08,0x44,0x12,0x01,0x10,0x48,0x10,0x10,0x00,0x00,0x00,
-    0x00,0x00,0x80,0x04,0x08,0x44,0x04,0x81,0x44,0x01,0x10,0x00,0x20,0x01,0x00,0x02,0x41,0x00,0x00,0x00,
-    0x00,0x00,0x02,0x10,0x42,0x11,0x20,0x00,0x11,0x20,0x01,0x48,0x89,0x24,0x20,0x88,0x00,0x08,0x08,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x20,0x04,0xa8,0x80,0x04,0x84,0x02,0x00,0x80,0x8a,0x00,0x09,0x20,0x40,0x00,
-    0x00,0x08,0x90,0x81,0x14,0x81,0x20,0x02,0x14,0x90,0x10,0x51,0x24,0x2a,0x00,0x54,0x80,0x01,0x00,0x00,
-    0x00,0x00,0x00,0x10,0x40,0x00,0x81,0x00,0x40,0x21,0x21,0x00,0x10,0x80,0x48,0x80,0x12,0x00,0x00,0x00,
-    0x00,0x20,0x04,0x04,0x15,0x24,0x14,0x12,0x85,0x44,0x04,0x0a,0x88,0x25,0x02,0x24,0x80,0x92,0x00,0x00,
-    0x00,0x01,0x20,0x41,0x00,0x12,0x91,0x48,0x28,0x11,0x42,0xa0,0x42,0x80,0x24,0x80,0x10,0x00,0x00,0x10,
-    0x00,0x00,0x00,0x10,0xa2,0xa0,0x44,0x22,0x45,0x04,0x28,0x51,0x14,0x11,0x00,0x2a,0x44,0x00,0x01,0x00,
-    0x00,0x00,0x05,0x04,0x11,0x09,0x20,0x88,0x48,0x42,0x82,0x84,0x42,0x4a,0xaa,0x20,0x11,0x24,0x04,0x00,
-    0x00,0x44,0x40,0x52,0x84,0x12,0x55,0x41,0x22,0x28,0x51,0x29,0x29,0x40,0x01,0x09,0x42,0x01,0x20,0x00,
-    0x01,0x00,0x09,0x04,0x11,0x49,0x01,0x14,0x8a,0x85,0x04,0x24,0x84,0x2a,0xd4,0x88,0x90,0xa8,0x00,0x00,
-    0x00,0x02,0x40,0xa1,0x44,0x08,0x44,0x84,0x21,0x58,0xa5,0x88,0x29,0x55,0x09,0x64,0x14,0x00,0x00,0x00,
-    0x00,0x08,0x08,0x0c,0x55,0x6d,0x55,0x33,0x14,0x25,0x14,0x52,0x89,0x00,0xc8,0x02,0x41,0x21,0x10,0x00,
-    0x00,0x00,0x82,0xa3,0x22,0xa2,0x91,0x54,0x8a,0x94,0x08,0x94,0x40,0xbe,0x25,0x6a,0x54,0x08,0x02,0x10,
-    0x00,0x82,0x21,0x10,0x8a,0x90,0xa9,0x0b,0x44,0xa9,0x52,0x92,0xbe,0x80,0xa9,0x29,0x21,0x00,0x00,0x00,
-    0x04,0x08,0x08,0x0d,0x29,0x5d,0x51,0x48,0xb3,0x45,0x25,0x6c,0xc0,0xd5,0x44,0xca,0x08,0x22,0x40,0x00,
-    0x00,0x00,0xa1,0x50,0x4a,0x20,0xaa,0xa2,0x88,0x25,0x29,0x03,0x2f,0x28,0x33,0x01,0x55,0x14,0x10,0x80,
-    0x00,0x12,0x09,0x29,0x2a,0xd6,0xa4,0xac,0xa7,0x92,0x85,0xac,0xa2,0x9b,0xaa,0x6c,0x80,0x40,0x02,0x00,
-    0x00,0x80,0x80,0x94,0x48,0x55,0x39,0x55,0x4a,0x38,0xb5,0x4b,0x10,0xa4,0x15,0x02,0x55,0x29,0x00,0x08,
-    0x04,0x02,0x2a,0x52,0xad,0x10,0x42,0xaa,0x92,0x95,0x08,0xa8,0xae,0xaa,0xa1,0x48,0x22,0x80,0x50,0x00,
-    0x00,0x10,0x05,0x15,0x52,0x93,0x1a,0x2a,0x15,0x21,0xb5,0x2c,0x29,0x4a,0x56,0x95,0x48,0x45,0x00,0x80,
-    0x00,0x41,0x42,0x92,0x25,0x4c,0x25,0xa5,0xda,0x51,0x42,0x83,0x25,0x29,0x51,0x54,0x11,0x20,0x10,0x00,
-    0x00,0x04,0x10,0x4a,0xb4,0x85,0x95,0x2a,0xa1,0xa5,0x08,0x62,0xd3,0x32,0xad,0x5a,0x6a,0x48,0x82,0x00,
-    0x01,0x00,0x8a,0x29,0x12,0xb2,0xa2,0x6a,0x95,0x48,0xdb,0x5c,0x0c,0x49,0x40,0xa5,0x05,0x22,0x08,0x10,
-    0x04,0x04,0x21,0x11,0xaa,0xaa,0x4d,0x54,0x55,0x03,0x50,0x42,0xf2,0xf4,0x7f,0x52,0x90,0x88,0x40,0x00,
-    0x00,0x11,0x04,0x8a,0x55,0x54,0xaa,0x4b,0x4a,0xa4,0xaf,0x2d,0x49,0x4b,0x55,0x12,0x95,0x21,0x00,0x40,
-    0x00,0x00,0xa4,0x55,0x55,0x2b,0x4d,0xa4,0xa5,0x17,0x20,0xa8,0x54,0x8a,0xa2,0x65,0x48,0x44,0x22,0x00,
-    0x00,0x22,0x01,0x09,0x48,0x0d,0x4a,0x53,0x19,0x28,0xaa,0xb6,0x15,0x25,0x15,0x24,0xa9,0x21,0x00,0x00,
-    0x08,0x00,0x52,0x92,0x51,0xa5,0x36,0xda,0xed,0x66,0x45,0x5b,0x5a,0xba,0xb4,0x8a,0x84,0x94,0x40,0x00,
-    0x01,0x04,0x80,0x51,0x2a,0x5a,0xa9,0x25,0x2c,0x99,0x32,0x14,0x95,0x4c,0x28,0x61,0x28,0x40,0x04,0x84,
-    0x00,0x11,0x12,0x8a,0xa4,0x45,0x22,0xa9,0x92,0xc5,0x49,0xcb,0x28,0x65,0xef,0x28,0x45,0x22,0x40,0x00,
-    0x00,0x00,0x20,0x54,0xa3,0x2b,0x52,0xac,0xa5,0x1b,0xa8,0x34,0xad,0xa2,0x15,0x22,0x20,0x88,0x08,0x00,
-    0x00,0x42,0x11,0x11,0x29,0xa4,0x14,0x53,0x52,0x61,0x13,0x80,0xa0,0xa9,0x6a,0x91,0x94,0x20,0x80,0x20,
-    0x09,0x00,0x84,0xa2,0x94,0xa1,0xa3,0x50,0x9a,0x2a,0xa0,0x57,0x1a,0xa9,0x02,0x44,0x11,0x02,0x20,0x80,
-    0x00,0x04,0x11,0x08,0x42,0x2a,0x54,0xaa,0xe2,0xa5,0x5a,0xaa,0xa8,0x84,0x80,0x12,0xa4,0xa4,0x00,0x00,
-    0x00,0x10,0x4a,0x25,0x50,0x94,0x92,0x2a,0x92,0x92,0x2a,0x51,0x12,0x6a,0x55,0xc8,0x08,0x00,0x84,0x00,
-    0x00,0x01,0x00,0x20,0x8a,0x21,0x24,0x85,0x14,0xa2,0x55,0x04,0x84,0x91,0x44,0x45,0x51,0x20,0x10,0x00,
-    0x00,0x80,0x08,0x0a,0x40,0x52,0x52,0x68,0x80,0x94,0xa2,0x52,0x6a,0x54,0x4a,0x12,0xa4,0x44,0x00,0x00,
-    0x00,0x04,0x41,0x20,0x8b,0x09,0x44,0x0b,0x55,0x2a,0x95,0x0d,0x08,0xa0,0x50,0xa9,0x00,0x08,0x80,0x80,
-    0x00,0x00,0x04,0x08,0x20,0x92,0x22,0xa0,0x4a,0x08,0x04,0x50,0x4a,0x0b,0x08,0x49,0x52,0x80,0x10,0x00,
-    0x00,0x11,0x00,0xa1,0x14,0x21,0x09,0x1a,0x20,0xaa,0x05,0x4d,0x09,0x20,0x45,0x02,0x08,0x12,0x40,0x00,
-    0x00,0x80,0x22,0x04,0x42,0x04,0x42,0x40,0x8a,0x11,0x50,0x10,0x40,0x49,0x20,0xd0,0xa0,0x40,0x01,0x00,
-    0x04,0x00,0x00,0x20,0x04,0xa2,0x95,0x25,0x20,0x82,0x0a,0x84,0x14,0x02,0x15,0x04,0x02,0x00,0x00,0x00,
-    0x00,0x00,0x10,0x04,0xa0,0x48,0x00,0x00,0x8a,0x28,0x80,0x49,0x42,0x98,0x80,0x21,0x20,0x88,0x80,0x00,
-    0x00,0x04,0x82,0x42,0x09,0x02,0xa9,0x52,0x20,0x82,0x25,0x00,0x10,0x02,0x29,0x08,0x04,0x20,0x00,0x00,
-    0x00,0x00,0x00,0x10,0x20,0x10,0x04,0x08,0x89,0x28,0x90,0x25,0x45,0x20,0x80,0x42,0x90,0x00,0x08,0x00,
-    0x00,0x00,0x01,0x02,0x45,0x00,0x50,0x00,0x22,0x00,0x02,0x80,0x10,0x0a,0x05,0x08,0x00,0x00,0x00,0x04,
-    0x00,0x00,0x08,0x00,0x00,0x42,0x02,0x4a,0x84,0x12,0x88,0x22,0x04,0x20,0x20,0x20,0x00,0x84,0x00,0x10,
-    0x00,0x20,0x40,0x10,0x88,0x10,0x10,0x00,0x10,0x80,0x20,0x00,0x80,0x81,0x04,0x84,0x92,0x10,0x40,0x40,
-    0x00,0x02,0x01,0x00,0x01,0x01,0x04,0x48,0x02,0x00,0x01,0x0a,0x10,0x14,0x08,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x08,0x48,0x08,0x41,0x00,0x80,0x08,0x84,0x00,0x00,0x40,0x40,0x20,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x41,0x00,0x00,0x08,0x04,0x20,0x00,0x10,0x20,0x21,0x01,0x02,0x04,0x10,0x00,0x00,0x00,
-    0x00,0x00,0x08,0x00,0x00,0x02,0x00,0x00,0x02,0x00,0x00,0x02,0x04,0x00,0x10,0x00,0x00,0x80,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x04,0x90,0x00,0x20,0x00,0x49,0x02,0x08,0x00,0x04,0x00,0x00,0x40,0x01,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x11,0x00,0x88,0x00,0x00,0x40,0x00,0x10,0x00,0x90,0x00,0x08,0x01,0x00,
-    0x00,0x00,0x00,0x02,0x20,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x20,0x00,0x02,0x01,0x00,0x00,0x00,0x00,
-    0x00,0x02,0x01,0x10,0x00,0x00,0x80,0x00,0x00,0x00,0x10,0x00,0x81,0x00,0x10,0x00,0x01,0x00,0x00,0x00,
-    0x00,0x00,0x04,0x00,0x00,0x20,0x00,0x08,0x00,0x40,0x00,0x80,0x00,0x00,0x00,0x00,0x04,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x80,0x02,0x00,0x00,0x00,0x00,0x04,0x00,0x08,0x00,0x00,0x00,0x00,0x00,0x00,
-
-};
-
-const lv_image_dsc_t noise_4 = {
-  .header = {
-    .magic = LV_IMAGE_HEADER_MAGIC,
-    .cf = LV_COLOR_FORMAT_I1,
-    .flags = 0,
-    .w = 160,
-    .h = 68,
-    .stride = 20,
-    .reserved_2 = 0,
-  },
-  .data_size = sizeof(noise_4_map),
-  .data = noise_4_map,
-  .reserved = NULL,
-};
-
