--- boards/shields/nice_view_elemental/src/images/noise_5.c
@@ -1,102 +1,0 @@
-#include <lvgl.h>
-
-#ifndef LV_ATTRIBUTE_MEM_ALIGN
-#define LV_ATTRIBUTE_MEM_ALIGN
-#endif
-
-#ifndef LV_ATTRIBUTE_NOISE_5
-#define LV_ATTRIBUTE_NOISE_5
-#endif
-
-static const
-LV_ATTRIBUTE_MEM_ALIGN LV_ATTRIBUTE_LARGE_CONST LV_ATTRIBUTE_NOISE_5
-uint8_t noise_5_map[] = {
-
-    0xff,0xff,0xff,0xff,0x00,0x00,0x00,0xff,
-
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x82,0x00,0x10,0x80,0x81,0x10,0x40,0x88,0x90,0x84,0x81,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x04,0x00,0x20,0x82,0x08,0x10,0x01,0x04,0x00,0x02,0x10,0x08,0x11,0x08,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x20,0x10,0x02,0x00,0x00,0x00,0x04,0x00,0x00,0x00,0x00,0x00,0x40,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x08,0x00,0x00,0x80,0x00,0x82,0x20,0x20,0x10,0x44,0x40,0x02,0x10,0x00,0x00,0x80,0x00,0x00,
-    0x00,0x00,0x01,0x01,0x04,0x12,0x48,0x10,0x84,0x80,0x42,0x11,0x10,0x84,0x00,0x08,0x44,0x08,0x00,0x00,
-    0x00,0x00,0x40,0x08,0x10,0x00,0x00,0x40,0x00,0x09,0x00,0x00,0x02,0x10,0xa2,0x42,0x00,0x00,0x40,0x00,
-    0x00,0x01,0x00,0x20,0x40,0x49,0x24,0x05,0x11,0x20,0x24,0x88,0x40,0x00,0x00,0x00,0x20,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x81,0x02,0x00,0x01,0x10,0x44,0x04,0x00,0x11,0x02,0x49,0x12,0xa2,0x02,0x41,0x00,0x00,
-    0x00,0x04,0x08,0x00,0x10,0x22,0x90,0x82,0x01,0x11,0x45,0x02,0x28,0x00,0x20,0x00,0x80,0x08,0x04,0x00,
-    0x00,0x10,0x01,0x04,0x41,0x08,0x08,0x48,0x54,0x40,0x28,0x48,0x84,0x51,0x09,0x10,0x24,0x20,0x00,0x00,
-    0x00,0x00,0x20,0x20,0x10,0xa1,0x20,0x21,0x00,0x09,0x04,0x02,0x22,0x02,0x50,0x45,0x01,0x00,0x20,0x00,
-    0x00,0x00,0x04,0x89,0x04,0x01,0x02,0x84,0x4a,0xa2,0x41,0x54,0x10,0xa9,0x04,0x90,0x28,0x04,0x80,0x40,
-    0x00,0x02,0x00,0x04,0xa2,0x14,0xa8,0x22,0x80,0x11,0x14,0x82,0x84,0x84,0x90,0x42,0x01,0x40,0x00,0x00,
-    0x00,0x40,0x49,0x22,0x14,0xa1,0x12,0x94,0x55,0x42,0x09,0x28,0x52,0x51,0x05,0x08,0xa4,0x10,0x00,0x00,
-    0x00,0x08,0x00,0x08,0x40,0x8a,0x50,0x4a,0x80,0x24,0x44,0x56,0x24,0x84,0xa0,0x44,0x10,0x02,0x21,0x00,
-    0x01,0x00,0x49,0x55,0x2a,0x55,0x4a,0x50,0x5a,0x49,0x14,0x20,0x94,0x2a,0x4d,0x41,0x81,0x40,0x00,0x00,
-    0x00,0x01,0x00,0x00,0x05,0x20,0xa4,0x02,0x21,0x24,0xaa,0x95,0x41,0x81,0x11,0x12,0x28,0x10,0x80,0x00,
-    0x00,0x04,0x15,0x54,0x68,0x54,0x95,0xa9,0x54,0x92,0x89,0x52,0xa5,0x36,0x54,0x84,0x85,0x04,0x08,0x04,
-    0x00,0x20,0x40,0x08,0x02,0x82,0x28,0x55,0x24,0x4a,0x22,0x28,0x0a,0x01,0x50,0xb1,0x50,0x50,0x00,0x80,
-    0x02,0x00,0x0a,0xa2,0xaa,0x51,0x47,0x0a,0x52,0x90,0x15,0x45,0x68,0xac,0x85,0x4a,0x0a,0x01,0x20,0x00,
-    0x00,0x08,0x40,0x48,0x08,0xba,0x14,0xb5,0x95,0x4e,0xd5,0x22,0x8a,0x23,0x24,0x91,0x44,0x94,0x02,0x00,
-    0x00,0x42,0x04,0x4a,0xa6,0x0a,0xab,0x25,0x44,0xa0,0x65,0x55,0x2b,0x4a,0x92,0xa5,0x12,0x20,0x48,0x00,
-    0x01,0x00,0x42,0x22,0x50,0xc0,0x82,0xa5,0x54,0xab,0x25,0x52,0x94,0x91,0x14,0x22,0xa0,0x88,0x80,0x00,
-    0x00,0x08,0x94,0x9a,0x2a,0xb7,0x2d,0x5a,0x92,0x94,0xab,0x28,0xd9,0x4c,0xd1,0x48,0x8a,0x42,0x00,0x20,
-    0x04,0x22,0x09,0x22,0x92,0x4a,0x8a,0x4a,0x55,0x6a,0xa0,0xab,0x4a,0xa5,0x2a,0x16,0x50,0x10,0x24,0x80,
-    0x00,0x00,0x40,0x4a,0xc1,0x94,0x69,0x4a,0xc9,0x14,0x95,0xad,0xa5,0x52,0xaa,0xe1,0x45,0x00,0x40,0x00,
-    0x00,0x80,0x12,0x25,0x2d,0x52,0xab,0x61,0x55,0x9a,0xd4,0x85,0x35,0xb8,0x55,0x32,0x32,0x52,0x00,0x00,
-    0x00,0x0a,0xa1,0x51,0x52,0x14,0x51,0x1d,0x54,0x8a,0x6b,0x13,0x92,0x86,0xc5,0x91,0x48,0x84,0x10,0x00,
-    0x00,0x00,0x08,0x95,0x4b,0xab,0x55,0xe7,0x25,0x35,0x4a,0xaa,0x4d,0x7b,0x50,0xa4,0x46,0x21,0x00,0x82,
-    0x04,0x82,0x42,0x4a,0xaa,0x28,0xaa,0x38,0xb3,0x8a,0xac,0x15,0xa2,0x89,0x5d,0x19,0x20,0x94,0x04,0x00,
-    0x00,0x10,0x04,0xa1,0x15,0x97,0x51,0x4a,0x95,0x53,0x49,0xa6,0xaa,0xec,0xa5,0x45,0x9a,0x40,0x50,0x00,
-    0x00,0x01,0x12,0x2d,0x22,0x6a,0xad,0x49,0x5c,0x46,0x9c,0x49,0x4a,0x94,0x5a,0x95,0x04,0x8a,0x00,0x00,
-    0x00,0x40,0x00,0x91,0x92,0x89,0x4a,0x56,0x87,0x15,0x62,0x4d,0x55,0x53,0x24,0x94,0xa2,0x20,0x00,0x20,
-    0x04,0x11,0x4a,0xaa,0x8a,0xd6,0xad,0x14,0xa9,0xdd,0xb6,0xa9,0x2a,0x69,0x0a,0xa2,0x51,0x4a,0x91,0x00,
-    0x00,0x00,0x08,0x54,0xa6,0x54,0x94,0x8a,0x62,0x6a,0x88,0x2d,0xb2,0x48,0xa9,0x1d,0x48,0x40,0x00,0x00,
-    0x00,0x85,0x22,0x22,0x55,0x22,0x34,0x81,0x1a,0xa5,0x72,0x96,0x8d,0xa5,0x55,0xa2,0x25,0x10,0xa0,0x00,
-    0x00,0x08,0x00,0x8d,0x2a,0x6a,0x95,0x6a,0xaa,0x52,0x1d,0xa8,0xe8,0x56,0xd2,0x51,0x88,0x4a,0x04,0x00,
-    0x00,0x00,0x2a,0x22,0x49,0x15,0x2a,0xa5,0x23,0x35,0x62,0x46,0x17,0x28,0x95,0xa6,0x55,0x00,0x10,0x10,
-    0x09,0x22,0x01,0x49,0x22,0x84,0xd0,0xaa,0xb9,0x55,0x10,0xa9,0xd2,0xca,0x90,0x91,0x00,0xa9,0x40,0x80,
-    0x00,0x00,0x92,0x21,0x18,0xb1,0x05,0x48,0x14,0x48,0x35,0x4a,0x11,0x21,0x54,0x44,0x95,0x00,0x00,0x00,
-    0x00,0x08,0x24,0x04,0x85,0x4a,0x40,0x24,0xcb,0x2d,0xa8,0x29,0x4e,0xad,0x25,0x21,0x40,0x22,0x00,0x00,
-    0x00,0x00,0x81,0x12,0x5a,0x91,0x1a,0x89,0x14,0x90,0xa3,0x28,0x20,0x22,0x52,0x48,0x2a,0x88,0x11,0x00,
-    0x00,0x40,0x08,0x41,0x40,0x49,0x44,0xa4,0x45,0x05,0x08,0xab,0x4d,0x54,0x82,0x8a,0x90,0x02,0x40,0x00,
-    0x00,0x02,0x41,0x14,0x8a,0x91,0x11,0x26,0x28,0x51,0x64,0x08,0x41,0x44,0x48,0x44,0x22,0x90,0x00,0x00,
-    0x00,0x08,0x10,0x02,0x11,0x28,0x44,0x91,0x8a,0x49,0x0a,0xa8,0xac,0x15,0x24,0xa1,0x10,0x02,0x00,0x00,
-    0x02,0x00,0x04,0x51,0x48,0x92,0xa2,0x54,0x50,0xa4,0x42,0x20,0x24,0x80,0x92,0x04,0x81,0x20,0x44,0x20,
-    0x00,0x00,0x40,0x08,0x22,0x04,0x50,0x81,0x25,0x09,0x10,0x89,0x49,0x29,0x49,0x51,0x24,0x08,0x00,0x00,
-    0x00,0x00,0x02,0xa2,0x28,0xa9,0x0a,0x28,0x88,0xa2,0xa9,0x12,0x40,0x40,0x24,0x08,0x01,0x01,0x00,0x00,
-    0x00,0x22,0x20,0x00,0x84,0x44,0x42,0x44,0x22,0x08,0x00,0x44,0xa2,0x25,0x40,0xa5,0x48,0x20,0x10,0x00,
-    0x00,0x00,0x00,0x8a,0x01,0x09,0x11,0x01,0x09,0x22,0xaa,0x08,0x08,0x90,0x08,0x00,0x02,0x02,0x00,0x00,
-    0x00,0x00,0x48,0x00,0x54,0x00,0x44,0x28,0x50,0x00,0x00,0xa0,0x80,0x44,0x82,0x92,0x50,0x00,0x00,0x00,
-    0x00,0x01,0x00,0x21,0x01,0x11,0x00,0x01,0x04,0xa4,0x8a,0x04,0x12,0x00,0x10,0x00,0x00,0x90,0x00,0x00,
-    0x00,0x80,0x01,0x04,0x04,0x04,0x49,0x08,0x50,0x00,0x20,0x80,0x40,0x94,0x82,0x22,0x08,0x00,0x00,0x80,
-    0x00,0x00,0x00,0x00,0x20,0x90,0x00,0x20,0x02,0x05,0x04,0x11,0x08,0x00,0x10,0x80,0x42,0x01,0x08,0x00,
-    0x00,0x00,0x08,0x10,0x82,0x01,0x08,0x01,0x10,0x90,0x00,0x00,0x02,0x10,0x00,0x00,0x00,0x04,0x00,0x00,
-    0x00,0x04,0x20,0x82,0x08,0x04,0x01,0x04,0x00,0x00,0x12,0x48,0x20,0x41,0x08,0x12,0x00,0x20,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x20,0x08,0x20,0x04,0x01,0x00,0x02,0x00,0x00,0x22,0x40,0x40,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x08,0x00,0x00,0x80,0x01,0x20,0x10,0x40,0x80,0x09,0x00,0x00,0x00,0x04,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x01,0x02,0x10,0x84,0x00,0x00,0x04,0x10,0x80,0x04,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x01,0x00,0x40,0x08,0x00,0x00,0x00,0x02,0x00,0x00,0x00,0x20,0x44,0x04,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x10,0x00,0x04,0x00,0x00,0x10,0x02,0x40,0x00,0x00,0x00,0x80,0x00,0x20,0x20,0x40,0x00,0x00,
-    0x00,0x01,0x00,0x00,0x00,0x00,0x02,0x00,0x00,0x00,0x20,0x80,0x10,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x08,0x00,0x00,0x08,0x00,0x08,0x00,0x00,0x04,0x00,0x00,0x00,0x00,0x00,0x00,0x08,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x20,0x00,0x20,0x00,0x00,0x00,0x00,0x00,0x80,0x00,0x80,0x00,0x00,0x00,
-
-};
-
-const lv_image_dsc_t noise_5 = {
-  .header = {
-    .magic = LV_IMAGE_HEADER_MAGIC,
-    .cf = LV_COLOR_FORMAT_I1,
-    .flags = 0,
-    .w = 160,
-    .h = 68,
-    .stride = 20,
-    .reserved_2 = 0,
-  },
-  .data_size = sizeof(noise_5_map),
-  .data = noise_5_map,
-  .reserved = NULL,
-};
-
