+++ boards/shields/nice_view_elemental/src/initialize_listeners.c
@@ -1,4 +1,4 @@
-#include "../../include/central/initialize_listeners.h"
+#include "../include/initialize_listeners.h"
 
 #include <limits.h>
 #include <lvgl.h>
@@ -15,11 +15,56 @@
 #include <zmk/events/endpoint_changed.h>
 #include <zmk/events/keycode_state_changed.h>
 #include <zmk/events/layer_state_changed.h>
+#include <zmk/events/split_peripheral_status_changed.h>
 #include <zmk/events/usb_conn_state_changed.h>
 #include <zmk/keymap.h>
+#include <zmk/split/bluetooth/peripheral.h>
 #include <zmk/usb.h>
-#include "../../include/main.h"
-#include "../../include/central/render.h"
+#include "../include/main.h"
+#include "../include/render.h"
+
+// #if IS_ENABLED(CONFIG_NICE_VIEW_ELEMENTAL_ANIMATION)
+// We want to pause the animation when the keyboard is idling.
+int activity_update_callback(const zmk_event_t* eh) {
+    struct zmk_activity_state_changed* ev = as_zmk_activity_state_changed(eh);
+    if (ev == NULL) {
+        return -ENOTSUP;
+    }
+
+    switch (ev->state) {
+        case ZMK_ACTIVITY_ACTIVE: {
+            start_animation();
+            // lv_timer_resume(timer);
+            break;
+        }
+        case ZMK_ACTIVITY_IDLE:
+        case ZMK_ACTIVITY_SLEEP: {
+            stop_animation();
+            // lv_timer_pause(timer);
+            break;
+        }
+        default: {
+            return -EINVAL;
+        }
+    }
+
+    return 0;
+}
+
+// Create a listener named `activity_update`. This name is then used to create a
+// subscription. When subscribed, `activity_update_callback` will be called.
+ZMK_LISTENER(
+    activity_update,
+    activity_update_callback
+);
+
+// Subscribe the `activity_update` listener to the `zmk_activity_state_changed`
+// event dispatched by ZMK.
+ZMK_SUBSCRIPTION(
+    activity_update,
+    zmk_activity_state_changed
+);
+// #endif
 
 struct states states;
 
@@ -76,6 +121,7 @@ static void connectivity_state_update_callback(struct connectivity_state state) 
 }
 
 static struct connectivity_state get_connectivity_state(const zmk_event_t* event) {
+#if (defined(CONFIG_ZMK_SPLIT) && defined(CONFIG_ZMK_SPLIT_ROLE_CENTRAL))
     const struct zmk_endpoint_instance selected_endpoint = zmk_endpoints_selected();
     const int active_profile_index = zmk_ble_active_profile_index();
     const bool active_profile_connected = zmk_ble_active_profile_is_connected();
@@ -87,6 +133,13 @@ static struct connectivity_state get_connectivity_state(const zmk_event_t* event
         .active_profile_connected = active_profile_connected,
         .active_profile_bonded = active_profile_bonded,
     };
+#else
+    const bool connected = zmk_split_bt_peripheral_is_connected();
+
+    struct connectivity_state state = {
+        .connected = connected,
+    };
+#endif
 
     return state;
 }
@@ -101,6 +154,7 @@ ZMK_DISPLAY_WIDGET_LISTENER(
     get_connectivity_state
 )
 
+#if (defined(CONFIG_ZMK_SPLIT) && defined(CONFIG_ZMK_SPLIT_ROLE_CENTRAL))
 // Subscribe the `widget_connectivity_state_update` listener to the
 // `zmk_endpoint_changed` event dispatched by ZMK.
 ZMK_SUBSCRIPTION(
@@ -123,7 +177,18 @@ ZMK_SUBSCRIPTION(
     // Triggered when the selected profile has changed.
     zmk_ble_active_profile_changed
 );
+#else
+// Subscribe the `widget_connectivity_state_update` listener to the
+// `zmk_split_peripheral_status_changed` event dispatched by ZMK.
+ZMK_SUBSCRIPTION(
+    widget_connectivity_state_update,
+    // Triggered when the peripheral was connected or disconnected from the
+    // central.
+    zmk_split_peripheral_status_changed
+);
+#endif
 
+#if (defined(CONFIG_ZMK_SPLIT) && defined(CONFIG_ZMK_SPLIT_ROLE_CENTRAL))
 static void layer_state_update_callback(struct layer_state state) {
     states.layer = state;
 
@@ -201,38 +266,15 @@ ZMK_SUBSCRIPTION(
     widget_modifiers_state_update,
     zmk_keycode_state_changed
 );
-
-// 1. Insert images here.
-LV_IMG_DECLARE(grid);
-LV_IMG_DECLARE(grid_inverse);
-const lv_image_dsc_t* images[] = {
-    &grid,
-    // &grid_inverse
-};
-
-static const unsigned int frame_count = sizeof(images) / sizeof(images[0]);
-
-void initialize_animation() {
-    lv_animimg_set_src(image_canvas, (const void**)images, frame_count);
-    // 2. Set the time for the whole animation.
-    lv_animimg_set_duration(image_canvas, 200);
-}
-
-void start_animation() {
-    lv_animimg_set_repeat_count(image_canvas, LV_ANIM_REPEAT_INFINITE);
-    lv_animimg_start(image_canvas);
-}
-
-void stop_animation() {
-    lv_animimg_set_repeat_count(image_canvas, 1);
-    lv_animimg_start(image_canvas);
-}
+#endif
 
 void initialize_listeners() {
+#if (defined(CONFIG_ZMK_SPLIT) && defined(CONFIG_ZMK_SPLIT_ROLE_CENTRAL))
     widget_layer_state_update_init();
+    widget_modifiers_state_update_init();
+#endif
     widget_connectivity_state_update_init();
     widget_battery_state_update_init();
-    widget_modifiers_state_update_init();
 
     initialize_animation();
     start_animation();
